+++ Pandora/Paradigm/Inventory/Optics.hs
@@ -1,10 +1,17 @@
 module Pandora.Paradigm.Inventory.Optics
 	(Lens, type (:-.), (|>), view, set, over, (^.), (.~), (%~)) where
 
-import Pandora.Core.Morphism ((.), ($))
-import Pandora.Paradigm.Inventory.Storage (Storage, access, position, retrofit)
+import Pandora.Core.Morphism ((.), ($), (!))
+import Pandora.Paradigm.Basis.Product (Product ((:*:)))
+import Pandora.Paradigm.Inventory.Stateful (Stateful (Stateful), statefully)
+import Pandora.Paradigm.Inventory.Storage (Storage (Storage), access, position, retrofit)
 import Pandora.Pattern.Functor.Covariant (Covariant ((<$)))
 import Pandora.Pattern.Functor.Extractable (Extractable (extract))
+import Pandora.Pattern.Functor.Adjoint (Adjoint ((-|), (|-)))
+
+instance Adjoint (Storage s) (Stateful s) where
+	v -| f = Stateful $ \s -> (:*:) s . f . Storage $ s :*: (v !)
+	Storage (s :*: f) |- g = extract . statefully s . g $ f s
 
 infixr 0 :-.
 type (:-.) src tgt = Lens src tgt