+++ CHANGELOG.md
@@ -104,3 +104,4 @@
 * Define `Representable` functor typeclass
 * Define `Pointable`, `Applicative`, `Bindable` and `Representable` instances for `->`
 * Define infix `Adjoint` operators - `-|` and `|-`
+* Define `Adjoint` instance for `Stateful` and `Storage` datatypes